tidy: Fix bugprone-unintended-char-ostream-output findings
Bug: 374984156
Test: m com.android.bt
Flag: EXEMPT PURE_REFACTOR

+++ system/audio_hal_interface/aidl/client_interface_aidl.cc
@@ -55,7 +55,7 @@ std::ostream& operator<<(std::ostream& os, const BluetoothAudioCtrlAck& ack) {
     case BluetoothAudioCtrlAck::FAILURE:
       return os << "FAILURE";
     default:
-      return os << "UNDEFINED " << static_cast<int8_t>(ack);
+      return os << "UNDEFINED " << static_cast<int>(ack);
   }
 }
